Update pytest configs for rename
