+++ pyproject.toml
@@ -34,10 +34,10 @@ dependencies = [
     "spherical-geometry >= 1.2.22",
     "stsci.imagestats >= 1.6.3",
     "drizzle >= 2.0.1",
-    "webbpsf >= 1.2.1",
-    # pysiaf required by webbpsf uses pkg_resources which doesn't
+    "stpsf >= 2.0.0",
+    # pysiaf required by stpsf uses pkg_resources which doesn't
     # exist in python 3.12 without setuptools.
-    # TODO: drop this dependency when webbpsf is dropped
+    # TODO: drop this dependency when stpsf is dropped
     # (or when pysiaf no longer requires pkg_resources).
     "setuptools >= 60 ; python_version >= '3.12'",
 ]
@@ -83,7 +83,7 @@ repository = "https://github.com/spacetelescope/romancal"
 romancal = "romancal.stpipe.integration:get_steps"
 
 [project.entry-points.pytest11]
-webbpsf = "pytest_romancal.webbpsf_plugin"
+stpsf = "pytest_romancal.stpsf_plugin"
 
 [project.scripts]
 roman_static_preview = "romancal.scripts.static_preview:command"
